+++ pyteal/ast/subroutine_test.py
@@ -1487,3 +1487,21 @@ def test_relaxed_abi_arg():
     )
 
     pt.compileTeal(program, mode=pt.Mode.Application, version=7)
+
+
+def test_override_abi_method_name():
+    def abi_meth(a: pt.abi.Uint64, b: pt.abi.Uint64, *, output: pt.abi.Uint64):
+        return output.set(a.get() + b.get())
+
+    mspec = ABIReturnSubroutine(abi_meth).method_spec().dictify()
+    assert mspec["name"] == "abi_meth"
+
+    mspec = ABIReturnSubroutine(abi_meth, overriding_name="add").method_spec().dictify()
+    assert mspec["name"] == "add"
+
+    @ABIReturnSubroutine.name_override("overriden_add")
+    def abi_meth_2(a: pt.abi.Uint64, b: pt.abi.Uint64, *, output: pt.abi.Uint64):
+        return output.set(a.get() + b.get())
+
+    mspec = abi_meth_2.method_spec().dictify()
+    assert mspec["name"] == "overriden_add"